requirements.txt
Django==3.2.16
djangorestframework==3.14.0
Pillow==9.5.0
six==1.16.0